feat: Correção controller

+++ app/Http/Controllers/MarcaController.php
@@ -7,13 +7,20 @@ use Illuminate\Http\Request;
 
 class MarcaController extends Controller
 {
+
+    protected $marca;
+    
+    public function __construct(Marca $marca){
+        $this->marca = $marca;
+    }
+
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
-        $marcas = Marca::all();
-        return $marcas;
+        $marcas = $this->marca->all();
+        return response()->json($marcas, 201);
     }
 
     /**
@@ -21,33 +28,49 @@ class MarcaController extends Controller
      */
     public function store(Request $request)
     {
-        $marca = Marca::create($request->all());
-        return $marca;
+        $marca = $this->marca->create($request->all());
+        return response()->json($marca, 200);
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(Marca $marca)
+    public function show(int $id)
     {
-        return $marca;
+        $marca = $this->marca->find($id);
+        
+        if($marca === null)
+        {
+            return response()->json(['erro' => 'Não econtrado!'], 404); 
+        }
+        return response()->json($marca, 200);
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Marca $marca)
+    public function update(Request $request,int $id)
     {
+        $marca = $this->marca->find($id);
+        if($marca === null)
+        {
+            return response()->json(['erro' => 'Não econtrado!'], 404); 
+        }
         $marca->update($request->all());
-        return $marca;
+        return response()->json($marca, 200);
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Marca $marca)
+    public function destroy(int $id)
     {
+        $marca = $this->marca->find($id);
+        if($marca === null)
+        {
+            return response()->json(['erro' => 'Não econtrado!'], 404); 
+        }
         $marca->delete();
-        return ['msg'=> 'A Marca foi removida com sucesso!'];
+        return response()->json(['msg'=> 'A Marca foi removida com sucesso!'], 200);
     }
 }
